inupmt_be/requirements.txt
Django
djangorestframework
drf_yasg
django-rest-swagger
django-cors-headers
Pillow
django-filter
